+++ packages/react-resizable-panels/src/utils/dom/getPanelElement.ts
@@ -1,8 +1,8 @@
 export function getPanelElement(
   id: string,
-  panelGroupElement: ParentNode | HTMLElement = document
+  scope: ParentNode | HTMLElement = document
 ): HTMLElement | null {
-  const element = panelGroupElement.querySelector(`[data-panel-id="${id}"]`);
+  const element = scope.querySelector(`[data-panel-id="${id}"]`);
   if (element) {
     return element as HTMLElement;
   }