+++ packages/react-resizable-panels/src/utils/dom/getPanelElementsForGroup.ts
@@ -1,10 +1,8 @@
 export function getPanelElementsForGroup(
   groupId: string,
-  panelGroupElement: ParentNode | HTMLElement = document
+  scope: ParentNode | HTMLElement = document
 ): HTMLElement[] {
   return Array.from(
-    panelGroupElement.querySelectorAll(
-      `[data-panel][data-panel-group-id="${groupId}"]`
-    )
+    scope.querySelectorAll(`[data-panel][data-panel-group-id="${groupId}"]`)
   );
 }